% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{add_log_to_name}
\alias{add_log_to_name}
\title{Add Log() to Variable Names}
\usage{
add_log_to_name(x, check = TRUE)
}
\arguments{
\item{x}{Character string representing the variable name}

\item{check}{Logical, if TRUE only adds log notation if 'log' or 'Log' not already present}
}
\value{
Character string with log transformation notation
}
\description{
Adds log transformation notation to variable names, removing existing 'log' or 'Log'
tags if present
}
