% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{cocopack_theme}
\alias{cocopack_theme}
\title{Create a Custom ggplot2 Theme}
\usage{
cocopack_theme(which = "default", text_size = 16)
}
\arguments{
\item{which}{Character string specifying the theme variant. Currently only 'default' is supported.}

\item{text_size}{Numeric value for the base text size in the theme. Default is 16.}
}
\value{
A ggplot2 theme object
}
\description{
Cocopack themes for ggplot2.
}
\details{
The default theme is a minimalist theme based on theme_bw() with gridlines removed.
}
