% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{flip_names}
\alias{flip_names}
\title{Flip Keys and Values in a Named Vector or List}
\usage{
flip_names(mapping)
}
\arguments{
\item{mapping}{A named vector or list where names will be swapped with values}
}
\value{
A named vector or list with keys and values flipped, preserving the original class
}
\description{
Takes a named vector or list and swaps the names with their values
}
