% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_package_bibtex}
\alias{get_package_bibtex}
\title{Get Package Citation in BibTeX Format}
\usage{
get_package_bibtex(x)
}
\arguments{
\item{x}{Package name}
}
\value{
Prints BibTeX citation
}
\description{
Get Package Citation in BibTeX Format
}
