% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{name_to_title}
\alias{name_to_title}
\title{Convert Names to Title Case}
\usage{
name_to_title(x)
}
\arguments{
\item{x}{Character string to convert to title case}
}
\value{
Character string in clean title case format
}
\description{
Converts variable names to clean title case format using janitor package
}
