% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{name_to_upper}
\alias{name_to_upper}
\title{Convert Names to Upper Case}
\usage{
name_to_upper(x)
}
\arguments{
\item{x}{Character string to convert to upper case}
}
\value{
Character string in clean upper case format
}
\description{
Converts variable names to clean upper case format using janitor package
}
