% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_funcs.R
\name{print_r_info}
\alias{print_r_info}
\title{Print system information about R}
\usage{
print_r_info()
}
\value{
Information about the R installation
}
\description{
Print system information about R
}
\examples{
print_r_info()
}
