% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{read_parquet_add_name}
\alias{read_parquet_add_name}
\title{Read Parquet and Add Filename Column}
\usage{
read_parquet_add_name(file)
}
\arguments{
\item{file}{Path to the Parquet file}
}
\value{
A tibble with an additional filename column
}
\description{
Reads a Parquet file and adds the filename as a column in the resulting dataframe
}
