% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{remove_log_label}
\alias{remove_log_label}
\title{Remove Log Labels from Column Names}
\usage{
remove_log_label(data)
}
\arguments{
\item{data}{A dataframe containing columns with log labels}
}
\value{
A dataframe with log labels removed from column names
}
\description{
Removes 'log_', 'log(', and ')' from column names in a dataframe
}
