% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_tools.R
\name{run_bootstrap}
\alias{run_bootstrap}
\title{Run Bootstrap Analysis}
\usage{
run_bootstrap(
  data_frame,
  group_vars,
  boot_fn,
  times = 1000,
  hypothesis = 0,
  parse = TRUE,
  simplify = TRUE
)
}
\arguments{
\item{data_frame}{Input data frame}

\item{group_vars}{Vector of grouping variables}

\item{boot_fn}{Bootstrap function to apply}

\item{times}{Number of bootstrap iterations}

\item{hypothesis}{Null hypothesis value}

\item{parse}{Whether to parse results into summary statistics}

\item{simplify}{Whether to return simplified results}
}
\value{
Data frame with bootstrap results
}
\description{
Performs bootstrap analysis on grouped data with progress bar
}
