% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_funcs.R
\name{say_hello}
\alias{say_hello}
\title{Say hello to someone (test function)}
\usage{
say_hello(name, exclamation = TRUE)
}
\arguments{
\item{name}{Name of a person}

\item{exclamation}{Whether to include an exclamation mark}
}
\description{
Say hello to someone (test function)
}
