% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_tools.R
\name{spearman_boot}
\alias{spearman_boot}
\title{Spearman Correlation with Bootstrap}
\usage{
spearman_boot(df, x_var, y_var, group_vars, R = 1000, parse = TRUE)
}
\arguments{
\item{df}{Input data frame}

\item{x_var}{X variable name}

\item{y_var}{Y variable name}

\item{group_vars}{Grouping variables}

\item{R}{Number of bootstrap iterations}

\item{parse}{Whether to parse results}
}
\value{
Bootstrap results for Spearman correlation
}
\description{
Spearman Correlation with Bootstrap
}
