% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_tools.R
\name{spearman_corr}
\alias{spearman_corr}
\title{Spearman Correlation Helper Function}
\usage{
spearman_corr(data, indices = NULL, x_var, y_var)
}
\arguments{
\item{data}{Data frame}

\item{indices}{Bootstrap indices}

\item{x_var}{X variable name}

\item{y_var}{Y variable name}
}
\value{
Correlation coefficient
}
\description{
Spearman Correlation Helper Function
}
\keyword{internal}
