% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{view_ggplot2_shapes}
\alias{view_ggplot2_shapes}
\title{Display Available ggplot2 Shape Palettes}
\usage{
view_ggplot2_shapes()
}
\value{
A ggplot object showing the shape palette
}
\description{
Creates a visualization showing all available shape options in ggplot2 (0-24)
with their corresponding numerical IDs.
}
