% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{vslice}
\alias{vslice}
\title{Slice a Vector}
\usage{
vslice(vector, start, stop = NULL)
}
\arguments{
\item{vector}{The input vector to slice}

\item{start}{Starting position of the slice}

\item{stop}{Optional ending position of the slice (defaults to vector length)}
}
\value{
A vector containing the selected slice
}
\description{
Returns a subset of a vector from start to stop position
}
